third_party/blink/web_tests/content_extraction/ruby-annotations-expected.txt

<Root> [box: 0,0 800x600]
  <Container> [box: 0,0 800x15]
    <Container> [box: 0,0 800x15]
      <Container> [box: 0,5 190x10]
        <Text> "     base1 " [box: 0,5 60x10]
        <Container> [box: 0,0 60x5]
          <Text> "anno1" [box: 0,0 60x5]
        <Text> "     base2 " [box: 60,5 70x10]
        <Container> [box: 60,0 70x5]
          <Text> "anno2" [box: 60,0 70x5]
        <Text> "     base3 " [box: 130,5 60x10]
        <Container> [box: 130,0 60x5]
          <Text> "anno3" [box: 130,0 60x5]
